Extract DB Client methods to builders
